requirements.txt
fastapi~=0.110.0
uvicorn~=0.28.0
python-dotenv~=1.0.0
pydantic~=2.11.4
pydantic-settings~=2.9.1
loguru~=0.7.3
requests~=2.32.3

llama-index~=0.12.35
llama-index-llms-google-genai~=0.1.12

redis~=6.1.0
orjson~=3.10.18